refactor: 중복되는 TodoItem 컴포넌트 삭제
